Lay foundation for Binary Clock
The only good binary clock app I can find on Android is so old that you
can't access its limited customization menu without a hardware menu
button. As such, I want to make my own, and I see this as a good excuse
to practice the basics of React Native in a simple app.

This commit is basically just the result of `react-native init` with
the TypeScript template, but I did figure out how to show the clock as
a sequence of periods that updates over time.

I have other homework to focus on now, but this has been a satisfying
start.