Respond to Screen Rotation
The Binary Clock now responds to the current device orientation.

When in portrait mode, the hours, minutes, and seconds will each be
displayed as a single column to better leverage the available vertical
space.

Landscape mode now shows the pairs of binary digits for each of hours,
minutes, and seconds. Previously I've obtained this view by locking my
phone in portrait mode then turning it on its side to make it look like
it was in landscape.