Add a brightness configuration setting for the Binary Clock
Someone using the Binary Clock by their bedside will likely prefer a
dimmer screen than a person using it at their office desk.

To accomodate both preferences, this commit adds a Settings screen with
a slider to adjust the clock's brightness. The settings are accessible
only via long pressing on the clock face in order to preserve the
minimalist feel of the Binary Clock (a Toast is shown on app launch to
inform the user of how to access the settings.)

Improves #1
Closes #2