build(ios): Sync MARKETING_VERSION with package.json
For iOS, the MARKETING_VERSION variable is used to set the semver
version of the app and must increment for each new release.

This commit instructs fastlane to pull the version from the
`package.json` -- just like I'm already doing for Android -- so that new
releases for Android and iOS only require updating the version number in
one place.