build: Make Android build numbers more consistent
The build number was being incremented when creating debug
builds, making it harder to roll back versions while switch between
branches.

Now, fastlane is reponsible for the version name and code, which
ensures that they are handled on every production build, and
are completed prior to the actual compilation of the app.