build: Require test file names to include `(spec|test)`
The default Jest config was considering my test fixture code as a
test suite because it was located under the `__tests__` directory.

Now, test files must end with `.test.ts` (or any of its common
variants. As a result, fixture code, which ends with `.ts` is no
longer considered a test suite.