fix: Reduce shakiness of `SettingRange`
SettingRange was invoking the `onValueChange` prop with
high frequency. When that function took less than nearly
instantaneous time to run, the slider would move around after
the user lifted his/her finger as it "caught up" with all the fired
events.

Now there's a debounce on that callback, which retains the
slider's reactivity, but dramatically reduces the lag.