perf: Remove multiple native dependencies
Removes the need for React Navigation by making the settings accessible by tapping the clock (which should be easier to find than a long press).

Additionally, removes the need for a dedicated save button -- preferring instead to auto save each setting.

And, the section of the screen rendering the clock now contains the appropriate onLayout handling to automatically compute whether the clock should render in portrait or landscape mode.