feat: Implement new Settings store
The current store has a lot of imperative usage, and I'd like a nicer abstraction that removes more concerns from the UI.

This new store is nice in that the API is similar to that of `useState` -- it returns a tuple of the current state and a function that facilitates easy state updates.

But the API abstracts away a lot of additional complexity, namely syncing with an external, persistent store.

---

There are a few other things that I like about this implementation.

(1) All settings are defined in a single `definitions.ts`.

It's nice to have one clear place to list every setting along with their defaults and logic for reading, validating, updating.

(2) Powerful, useful TypeScript typing.

Newly added settings to `definitions.ts` are automatically reflected in the type signature of `useSettings`

(3) Clear separation of concerns

The `NamedSubscriberStore` provides all the primitives to connect `useSyncExternalStore` with any store that can be modelled as a `Record<string, any>`. That separates all of the "boilerplatey" subscriber/notifier stuff from the more specific logic of the `SettingsStore`. And it's also nice that `SettingsStore` doesn't really care how the settings are persisted (if at all). Even the `definitions.ts` accesses its persistence layer through the `KeyValueStore` which is a nice abstraction that handles debouncing multiple writes to the same setting.

All these efforts will make it easier to extract this logic into a separate library if desired, or to swap out the persistence layer (likely once I try to add Windows/MacOS/Web support).