refactor: Extract components from `BinaryClockScreen`
The prior implementation of the settings store required everything close together on the screen to enable coordinating values between the settings pane and the clock.

Now that the new store can keep settings in sync even when accessed from different places, we can do a better separation of concerns.

Now the `BinaryClockScreen` is only responsible for coordinating the relative display sizes of the clock vs. the settings. Each of those concerns now has its own component.

Additionally, this change now means the settings components are no longer "controlled" by the clock screen, which fixes a UI jumping bug that could occur when moving sliders back and forth really quickly.