feat: Make "Show hints" the first setting
One of the most common topics in app reviews is people asking how to read a binary clock.

When this setting was at the end of the list, it was more likely to be "below the scroll" when opening the settings, and therefore unnoticed.

Now, it will always be at the top, and the caption is more targeted towards explaining the purpose of the setting's existence.

I think this strikes a nice balance between a clean clock as the default, but easily accessible learning opportunities for newcomers.

closes #6